Update BigDataCloud client usage and API models
- Reorganize and expand API usage examples in Program.cs
- Add new examples: Network by IP, ASN Info (short), ASN Rank List, Tor Exit Nodes
- Use more realistic test data and IPv6 in geolocation
- Remove ReverseGeocodeToCityAsync; focus on richer geocoding
- Correct full geolocation endpoint and GraphQL URL formatting
- Make ValidatePhoneByIpAsync accept optional IP address
- Change asnNumeric fields from int to long for ASN support
- Update Location.timeZone to use TimezoneResponse object
- Refine query builders for nested country info and currency code only
- Modernize code to align with latest BigDataCloud API features

+++ src/BigDataCloud/Models/Network.cs
@@ -82,7 +82,7 @@ public class AsnInfoResponse
     [JsonPropertyName("asn")]                    public string? Asn { get; set; }
 
     /// <summary>The Autonomous System Number as an unsigned integer (e.g. 13335).</summary>
-    [JsonPropertyName("asnNumeric")]             public int AsnNumeric { get; set; }
+    [JsonPropertyName("asnNumeric")]             public long AsnNumeric { get; set; }
 
     /// <summary>The organisation or entity the AS is registered for, as recorded in the RIR database.</summary>
     [JsonPropertyName("organisation")]           public string? Organisation { get; set; }
